Revert "Add Typescript support and partial conversion (#54)"
This reverts commit 8626d172c904e5b98481fa3c6683d094b4325d04.
