+++ .gitignore
@@ -57,5 +57,3 @@ typings/
 # dotenv environment variables file
 .env
 
-# Typescript cache
-tsconfig.tsbuildinfo
