--- .mocharc.yaml
@@ -1,5 +1,0 @@
-extension: [js, ts]
-require: 'ts-node/register'
-spec:
-  - 'tests/**.test.js'
-  - 'tests/**.test.ts'
