+++ package.json
@@ -11,11 +11,10 @@
     "yarn": "1.x"
   },
   "scripts": {
-    "start": "ts-node -r dotenv/config index.ts",
-    "lint": "tsc --project . --noemit && eslint . --ext .js --ext .ts",
-    "fix": "eslint . --fix --ext .js --ext .ts",
-    "format": "prettier --check ./.eslintrc.js ./tests/.eslintrc.json **/*",
-    "test": "mocha",
+    "start": "node -r dotenv/config index.js",
+    "lint": "eslint . --ext .js",
+    "format": "prettier ./.eslintrc.json -c **/*",
+    "test": "mocha tests/**.test.js",
     "coverage": "nyc npm run test"
   },
   "dependencies": {
@@ -27,11 +26,8 @@
     "sinon-chai": "^3.5.0"
   },
   "devDependencies": {
-    "@types/node": "^14.14.41",
-    "@typescript-eslint/eslint-plugin": "^4.22.0",
-    "@typescript-eslint/parser": "^4.22.0",
     "chai": "^4.2.0",
-    "eslint": "^7.25.0",
+    "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
     "faker": "^5.1.0",
@@ -40,9 +36,7 @@
     "mocha": "^8.1.3",
     "nyc": "^15.1.0",
     "prettier": "^2.0.5",
-    "snowflake-util": "^1.0.1",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.2.4"
+    "snowflake-util": "^1.0.1"
   },
   "husky": {
     "hooks": {
