--- tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    // From https://github.com/microsoft/TypeScript/wiki/Node-Target-Mapping
-    "lib": ["ES2020"],
-    "module": "CommonJS",
-    "target": "ES2020",
-
-    // Other
-    "incremental": true,
-    "strict": true,
-    "noImplicitAny": true,
-    "alwaysStrict": true,
-    "checkJs": false,
-    "esModuleInterop": true,
-  }
-}
